requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
Pillow==9.5.0
requests==2.31.0
urllib3==2.0.2
